requirements.txt
flask==3.1.2
flask_sqlalchemy==3.1.1
gunicorn==23.0.0
openai==2.1.0
cloudinary==1.44.1
requests==2.32.5
python-dotenv==1.0.1
pillow==11.0.0
